requirements.txt
python-dotenv
langchain-openai
langchain-core
pandas
streamlit
langchain
langchain-community
langgraph
langgraph-supervisor
pip install hf_xet
burdoc
ipywidgets